Unique PolicyForm components for Global Policies
