+++ portals/publisher/src/main/webapp/source/src/app/components/GlobalPolicies/Policies/PolicyForm/SourceDetails.tsx
@@ -16,27 +16,41 @@
 * under the License.
 */
 
-import React, { FC, useContext } from 'react';
+import React, { FC } from 'react';
+import { makeStyles, Theme } from '@material-ui/core';
+import Box from '@material-ui/core/Box';
+import FormControlLabel from '@material-ui/core/FormControlLabel';
+import FormGroup from '@material-ui/core/FormGroup';
+import Checkbox from '@material-ui/core/Checkbox';
+import Typography from '@material-ui/core/Typography';
 import { FormattedMessage } from 'react-intl';
+import FormControl from '@material-ui/core/FormControl';
+import CloudDownloadIcon from '@material-ui/icons/CloudDownload';
+import Button from '@material-ui/core/Button';
 import Utils from 'AppData/Utils';
 import API from 'AppData/api.js';
 import { Alert } from 'AppComponents/Shared';
 import CONSTS from 'AppData/Constants';
-import ApiContext from 'AppComponents/Apis/Details/components/ApiContext';
-import SourceDetailsShared from 'AppComponents/Shared/PoliciesUI/SourceDetails';
-import { ACTIONS } from './PolicyCreateForm';
-import UploadPolicyDropzone from './UploadPolicyDropzone';
+
+const useStyles = makeStyles((theme: Theme) => ({
+    mandatoryStar: {
+        color: theme.palette.error.main,
+        marginLeft: theme.spacing(0.1),
+    },
+    formGroup: {
+        display: 'flex',
+        flexDirection: 'row',
+    },
+}));
+
+export const GATEWAY_TYPE_LABELS = {
+    SYNAPSE: 'Regular Gateway',
+    CC: 'Choreo Connect'
+}
 
 interface SourceDetailsProps {
     supportedGateways: string[];
-    synapsePolicyDefinitionFile?: any[];
-    setSynapsePolicyDefinitionFile?: React.Dispatch<React.SetStateAction<any[]>>;
-    ccPolicyDefinitionFile?: any[];
-    setCcPolicyDefinitionFile?: React.Dispatch<React.SetStateAction<any[]>>;
-    dispatch?: React.Dispatch<any>;
-    isViewMode?: boolean;
     policyId?: string;
-    isAPISpecific?: boolean;
 }
 
 /**
@@ -46,94 +60,156 @@ interface SourceDetailsProps {
  */
 const SourceDetails: FC<SourceDetailsProps> = ({
     supportedGateways,
-    synapsePolicyDefinitionFile,
-    setSynapsePolicyDefinitionFile,
-    ccPolicyDefinitionFile,
-    setCcPolicyDefinitionFile,
-    dispatch,
-    isViewMode,
     policyId,
-    isAPISpecific,
 }) => {
-    const { api } = useContext<any>(ApiContext);
-
-    // Validates whether atleast one gateway type (i.e. synapse, or CC ) is selected
-    // True if none of the available gateways are selected.
-    const supportedGatewaysError = supportedGateways.length === 0;
-
-    /**
-     * Function to handle supported gateways related checkbox changes
-     * @param {React.ChangeEvent<HTMLInputElement>} event event
-     */
-    const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-        if (dispatch) {
-            dispatch({
-                type: ACTIONS.UPDATE_SUPPORTED_GATEWAYS,
-                name:
-                    event.target.name === 'regularGateway'
-                        ? CONSTS.GATEWAY_TYPE.synapse
-                        : CONSTS.GATEWAY_TYPE.choreoConnect,
-                checked: event.target.checked,
-            });
-        }
-    };
+    const classes = useStyles();
 
     /**
      * Hanlde policy download
      */
     const handlePolicyDownload = () => {
-        if (policyId) {
-            if (isAPISpecific) {
-                const globalPolicyContentPromise = API.getOperationPolicyContent(
-                    policyId,
-                    api.id,
-                );
-                globalPolicyContentPromise
-                    .then((globalPolicyResponse) => {
-                        Utils.forceDownload(globalPolicyResponse);
-                    })
-                    .catch((error) => {
-                        console.error(error);
-                        Alert.error(
-                            <FormattedMessage
-                                id='Apis.Details.Policies.PolicyForm.SourceDetails.apiSpecificPolicy.download.error'
-                                defaultMessage='Something went wrong while downloading the policy'
-                            />,
-                        );
-                    });
-            } else {
-                const commonPolicyContentPromise =
-                    API.getCommonOperationPolicyContent(policyId);
-                commonPolicyContentPromise
-                    .then((commonPolicyResponse) => {
-                        Utils.forceDownload(commonPolicyResponse);
-                    })
-                    .catch((error) => {
-                        console.error(error);
-                        Alert.error(
-                            <FormattedMessage
-                                id='Apis.Details.Policies.PolicyForm.SourceDetails.commonPolicy.download.error'
-                                defaultMessage='Something went wrong while downloading the policy'
-                            />,
-                        );
-                    });
-            }
+        if (policyId) { 
+            const commonPolicyContentPromise =
+                API.getCommonOperationPolicyContent(policyId);
+            commonPolicyContentPromise
+                .then((commonPolicyResponse) => {
+                    Utils.forceDownload(commonPolicyResponse);
+                })
+                .catch((error) => {
+                    console.error(error);
+                    Alert.error(
+                        <FormattedMessage
+                            id='Apis.Details.Policies.PolicyForm.SourceDetails.commonPolicy.download.error'
+                            defaultMessage='Something went wrong while downloading the policy'
+                        />,
+                    );
+                });   
         }
     };
 
-    return(
-        <SourceDetailsShared
-            supportedGateways={supportedGateways}
-            synapsePolicyDefinitionFile={synapsePolicyDefinitionFile}
-            setSynapsePolicyDefinitionFile={setSynapsePolicyDefinitionFile}
-            ccPolicyDefinitionFile={ccPolicyDefinitionFile}
-            setCcPolicyDefinitionFile={setCcPolicyDefinitionFile}
-            isViewMode={isViewMode}
-            handlePolicyDownload={handlePolicyDownload}
-            supportedGatewaysError={supportedGatewaysError}
-            handleChange={handleChange}
-            UploadPolicyDropzone={UploadPolicyDropzone}
-        />
+    return (
+        <Box display='flex' flexDirection='row' mt={1} data-testid='gateway-details-panel'>
+            <Box width='40%' pt={3} mb={2}>
+                <Box width='90%'>
+                    <Typography
+                        color='inherit'
+                        variant='subtitle2'
+                        component='div'
+                    >
+                        <FormattedMessage
+                            id='Apis.Details.Policies.PolicyForm.SourceDetails.title'
+                            defaultMessage='Gateway Specific Details'
+                        />
+                    </Typography>
+                    <Typography color='inherit' variant='caption' component='p'>
+                        <FormattedMessage
+                            id='Apis.Details.Policies.PolicyForm.SourceDetails.description'
+                            defaultMessage={
+                                'Define the Gateway (s) that will be supporting this policy. ' +
+                                'Based off of this selection, you can upload the relevant business ' +
+                                'logic inclusive policy file.'
+                            }
+                        />
+                    </Typography>
+                </Box>
+            </Box>
+            <Box width='60%'>
+                <Box display='flex' flexDirection='row' alignItems='center' data-testid='supported-gateways-form'>
+                    <Typography color='inherit' variant='body1' component='div'>
+                        <FormattedMessage
+                            id='Apis.Details.Policies.PolicyForm.SourceDetails.form.supported.gateways.label'
+                            defaultMessage='Supported Gateways'
+                        />
+                        <sup className={classes.mandatoryStar}>*</sup>
+                    </Typography>
+                    <Box
+                        flex='1'
+                        display='flex'
+                        flexDirection='row-reverse'
+                        justifyContent='space-around'
+                    >
+                        <FormControl
+                            required
+                            component='fieldset'
+                            variant='standard'
+                            margin='normal'
+                        >
+                            <FormGroup className={classes.formGroup}>
+                                <FormControlLabel
+                                    control={
+                                        <Checkbox
+                                            name='regularGateway'
+                                            color='primary'
+                                            checked={supportedGateways.includes(
+                                                CONSTS.GATEWAY_TYPE.synapse,
+                                            )}
+                                        />
+                                    }
+                                    label={GATEWAY_TYPE_LABELS.SYNAPSE}
+                                    data-testid='regular-gateway-label'
+                                />
+                                <FormControlLabel
+                                    control={
+                                        <Checkbox
+                                            name='choreoConnect'
+                                            color='primary'
+                                            checked={supportedGateways.includes(
+                                                CONSTS.GATEWAY_TYPE.choreoConnect,
+                                            )}
+                                        />
+                                    }
+                                    label={GATEWAY_TYPE_LABELS.CC}
+                                    data-testid='choreo-connect-label'
+                                />
+                            </FormGroup>
+                        </FormControl>
+                    </Box>
+                </Box>
+
+                <Box display='flex' flexDirection='row' alignItems='center'>
+                    <Typography
+                        color='inherit'
+                        variant='subtitle2'
+                        component='div'
+                    >
+                        <FormattedMessage
+                            id='Apis.Details.Policies.PolicyForm.SourceDetails.form.policy.file.title'
+                            defaultMessage='Policy File(s)'
+                        />
+                        <sup className={classes.mandatoryStar}>*</sup>
+                    </Typography>
+                </Box>
+                <Typography color='inherit' variant='caption' component='p'>
+                    <FormattedMessage
+                        id='Apis.Details.Policies.PolicyForm.SourceDetails.form.policy.file.description'
+                        defaultMessage='Policy file contains the business logic of the policy'
+                    />
+                </Typography>
+                <Box
+                    flex='1'
+                    display='flex'
+                    flexDirection='row'
+                    justifyContent='left'
+                    mt={3}
+                    mb={3}
+                >
+                    <Button
+                        aria-label='download-policy'
+                        variant='contained'
+                        data-testid='download-policy-file'
+                        size='large'
+                        color='primary'
+                        onClick={handlePolicyDownload}
+                        endIcon={<CloudDownloadIcon />}
+                    >
+                        <FormattedMessage
+                            id='Apis.Details.Policies.PolicyForm.SourceDetails.form.policy.file.download'
+                            defaultMessage='Download Policy'
+                        />
+                    </Button>
+                </Box>
+            </Box>
+        </Box>
     );
 };
